백준

+++ codingTest/src/level1/N_11382.java
@@ -6,11 +6,11 @@ public class N_11382 {
 	public static void main(String[] args) {
 		Scanner sc = new Scanner(System.in);
 		
-		int a,b,c;
+		long a,b,c;
 		
-		a = sc.nextInt();
-		b = sc.nextInt();
-		c = sc.nextInt();
+		a = sc.nextLong();
+		b = sc.nextLong();
+		c = sc.nextLong();
 		
 		System.out.printf("%d", a+b+c);
